Refactor: Consolidate API endpoints into license-api
